Changed margin datatype to decimal instead of int

+++ HotelPriceScout/Data/Function/IDataScraper.cs
@@ -6,6 +6,6 @@ namespace HotelPriceScout.Data.Function
     public interface IDataScraper
     {
         public BookingSite BookingSite { get; }
-        void StartScraping(int margin);
+        void StartScraping(decimal margin);
     }
 }
